refactor: split engine/review.rs into focused modules
- Extract auto_merge_pr + handle_review_changes + helpers to engine/auto_merge.rs
- Extract review_open_prs to engine/review_poll.rs
- Add dedup_reviews() shared helper (replaces duplicated inline dedup logic)
- Add attribution_footer() helper (replaces 6 identical inline footer formats)
- Extract ensure_pr_exists() from review_and_merge() to reduce nesting depth
- Move tests for moved functions to their respective modules

review.rs: 2349 → 1022 lines (56% reduction)
All 882 tests pass, cargo fmt + clippy clean.

Resolves #914
