+++ src/engine/auto_merge.rs
@@ -1,0 +1,790 @@
+//! PR auto-merge and review-changes handling.
+//!
+//! Extracted from `engine/review.rs`. Handles:
+//! - [`auto_merge_pr`]: merge an approved PR after CI passes
+//! - [`handle_review_changes`]: re-dispatch agent to address review feedback
+//! - [`dedup_reviews`]: deduplicate GitHub reviews by reviewer (keep latest)
+//! - [`any_changes_requested_in_reviews`]: check if any reviewer's latest review blocks merge
+//! - [`attribution_footer`]: standard "Commented/Reviewed by bot" footer for GitHub comments
+
+use crate::backends::{ExternalBackend, ExternalTask, Status};
+use crate::engine::cleanup::{
+    cleanup_task_worktree, store_increment, store_reset_failure_counters, store_set,
+};
+use crate::engine::tasks::TaskManager;
+use crate::github::http::GhHttp;
+use crate::github::types::GitHubReview;
+use crate::store::TaskStore;
+use crate::{config, engine};
+use std::collections::HashMap;
+use std::sync::{Arc, OnceLock};
+use tokio::sync::Semaphore;
+
+/// Maximum number of times we will attempt to rebase and retry a conflicting PR
+/// before giving up and blocking the task for human intervention.
+/// Both `review_open_prs` and `auto_merge_pr` use this constant so the limit is
+/// always consistent regardless of which code path increments the counter first.
+pub(crate) const MAX_MERGE_CONFLICT_RETRIES: u64 = 3;
+
+/// Maximum number of CI failure or CI timeout events in `auto_merge_pr` before
+/// the task is blocked for human intervention instead of re-entering NeedsReview.
+const MAX_CI_MERGE_FAILURES: u64 = 3;
+
+// Global semaphore limiting concurrent CI polling loops across all tasks.
+static CI_POLL_SEMAPHORE: OnceLock<Arc<Semaphore>> = OnceLock::new();
+
+fn ci_poll_semaphore() -> &'static Arc<Semaphore> {
+    CI_POLL_SEMAPHORE.get_or_init(|| Arc::new(Semaphore::new(3)))
+}
+
+/// Build a standard attribution footer for GitHub comments posted by orch bots.
+///
+/// `verb` is "Reviewed" or "Commented" depending on context.
+pub(crate) fn attribution_footer(verb: &str, agent: &str, model: &str) -> String {
+    format!(
+        "\n\n---\n*{} {}[bot] via [Orch](https://github.com/gabrielkoerich/orch) using `{}`*",
+        verb, agent, model
+    )
+}
+
+/// Deduplicate GitHub reviews by reviewer, keeping only the latest per reviewer.
+///
+/// GitHub returns reviews in chronological order. COMMENTED and DISMISSED reviews
+/// are ignored — they carry no approval/rejection signal.
+pub(crate) fn dedup_reviews<'a>(reviews: &'a [GitHubReview]) -> HashMap<String, &'a GitHubReview> {
+    let mut by_reviewer: HashMap<String, &'a GitHubReview> = HashMap::new();
+    for review in reviews {
+        if review.state != "APPROVED" && review.state != "CHANGES_REQUESTED" {
+            continue;
+        }
+        let dominated = by_reviewer
+            .get(&review.user.login)
+            .is_none_or(|prev| review.submitted_at > prev.submitted_at);
+        if dominated {
+            by_reviewer.insert(review.user.login.clone(), review);
+        }
+    }
+    by_reviewer
+}
+
+/// Returns `true` if any reviewer's **latest** review is `CHANGES_REQUESTED`.
+pub(crate) fn any_changes_requested_in_reviews(reviews: &[GitHubReview]) -> bool {
+    dedup_reviews(reviews)
+        .values()
+        .any(|r| r.state == "CHANGES_REQUESTED")
+}
+
+/// Auto-merge a PR after review approval.
+///
+/// Checks that the automated review comment says "approve" and that CI checks
+/// are green before merging. If CI fails, sets task back to routed so the
+/// agent is re-dispatched to fix the issues.
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn auto_merge_pr(
+    task: &ExternalTask,
+    branch: &str,
+    _backend: &Arc<dyn ExternalBackend>,
+    repo: &str,
+    review_agent: &str,
+    review_model: &str,
+    task_manager: &Arc<TaskManager>,
+    store: &Arc<TaskStore>,
+) -> anyhow::Result<()> {
+    // 1. Get PR number from branch
+    let gh = GhHttp::new()?;
+    let pr_number = match gh.get_pr_number(repo, branch).await? {
+        Some(n) => n,
+        None => {
+            anyhow::bail!("no open PR found for branch {}", branch);
+        }
+    };
+
+    // 2. Verify the automated review comment says "approve"
+    let review_status = gh.get_automated_review_status(repo, pr_number).await?;
+    match review_status.as_deref() {
+        Some("approve") => {
+            tracing::info!(task_id = task.id.0, pr_number, "automated review approved");
+        }
+        Some("changes_requested") => {
+            tracing::warn!(
+                task_id = task.id.0,
+                pr_number,
+                "automated review says changes_requested — approve comment was not posted, returning error"
+            );
+            anyhow::bail!(
+                "approve comment was not posted (latest review is changes_requested); task should be re-queued"
+            );
+        }
+        _ => {
+            tracing::info!(
+                task_id = task.id.0,
+                pr_number,
+                "no automated review comment found — proceeding with merge"
+            );
+        }
+    }
+
+    // 3. Rerun the failed push-triggered review gate check so the commit status flips green
+    if let Err(e) = gh
+        .rerun_failed_workflow(repo, "Orch Review Gate", branch)
+        .await
+    {
+        tracing::debug!(
+            task_id = task.id.0,
+            error = %e,
+            "failed to rerun orch-review workflow (may not have a failed run)"
+        );
+    }
+
+    // 4. Wait for CI checks to pass (poll up to 5 minutes)
+    let max_wait = std::time::Duration::from_secs(300);
+    let poll_interval = std::time::Duration::from_secs(15);
+    let start = std::time::Instant::now();
+
+    // Acquire a global permit to limit concurrent CI polling loops across tasks.
+    let _permit = ci_poll_semaphore().clone().acquire_owned().await;
+
+    loop {
+        let (state, total, passing, failing, pending) =
+            gh.get_combined_status(repo, branch).await?;
+
+        tracing::info!(
+            task_id = task.id.0,
+            pr_number,
+            state = %state,
+            total,
+            passing,
+            failing,
+            pending,
+            "CI status check"
+        );
+
+        match state.as_str() {
+            "success" => break,
+            "failure" => {
+                let ci_failures = store_increment(
+                    &Some(Arc::clone(store)),
+                    repo,
+                    &task.id.0,
+                    "ci_merge_failures",
+                )
+                .await;
+                if ci_failures >= MAX_CI_MERGE_FAILURES {
+                    tracing::error!(
+                        task_id = task.id.0,
+                        pr_number,
+                        ci_failures,
+                        "CI failure limit reached — blocking for human intervention"
+                    );
+                    task_manager
+                        .update_task_status(&task.id, Status::Blocked)
+                        .await?;
+                } else {
+                    tracing::warn!(
+                        task_id = task.id.0,
+                        pr_number,
+                        failing,
+                        ci_failures,
+                        "CI failed — re-routing to agent to fix"
+                    );
+                    task_manager
+                        .update_task_status(&task.id, Status::Routed)
+                        .await?;
+                    store_reset_failure_counters(&Some(Arc::clone(store)), repo, &task.id.0).await;
+                }
+                return Ok(());
+            }
+            _ => {
+                // pending — wait up to max_wait
+                if start.elapsed() >= max_wait {
+                    let ci_failures = store_increment(
+                        &Some(Arc::clone(store)),
+                        repo,
+                        &task.id.0,
+                        "ci_merge_failures",
+                    )
+                    .await;
+                    if ci_failures >= MAX_CI_MERGE_FAILURES {
+                        tracing::error!(
+                            task_id = task.id.0,
+                            ci_failures,
+                            "CI timeout limit reached — blocking for human intervention"
+                        );
+                        task_manager
+                            .update_task_status(&task.id, Status::Blocked)
+                            .await?;
+                    } else {
+                        tracing::warn!(
+                            task_id = task.id.0,
+                            ci_failures,
+                            "CI checks still pending after timeout — re-routing to agent"
+                        );
+                        task_manager
+                            .update_task_status(&task.id, Status::Routed)
+                            .await?;
+                        store_reset_failure_counters(&Some(Arc::clone(store)), repo, &task.id.0)
+                            .await;
+                    }
+                    return Ok(());
+                }
+            }
+        }
+
+        tokio::time::sleep(poll_interval).await;
+    }
+
+    // 5. Re-verify no reviewer has requested changes since the CI wait started.
+    let post_ci_reviews = gh.get_pr_reviews(repo, pr_number).await?;
+    if any_changes_requested_in_reviews(&post_ci_reviews) {
+        tracing::warn!(
+            task_id = task.id.0,
+            pr_number,
+            "a reviewer requested changes during CI wait — aborting merge"
+        );
+        anyhow::bail!("a reviewer requested changes during CI wait — aborting merge");
+    }
+
+    tracing::info!(
+        task_id = task.id.0,
+        pr_number,
+        branch = %branch,
+        "merging PR"
+    );
+
+    // 6. Merge via gh CLI
+    if let Err(e) = gh.merge_pr(repo, pr_number, true).await {
+        let err_msg = e.to_string().to_lowercase();
+        let is_conflict = err_msg.contains("405")
+            || err_msg.contains("not mergeable")
+            || err_msg.contains("merge conflict");
+
+        if is_conflict {
+            let retries =
+                engine::cleanup::store_get_field(store, repo, &task.id.0, "merge_conflict_retries")
+                    .await
+                    .and_then(|s| s.parse::<u64>().ok())
+                    .unwrap_or(0);
+            if retries >= MAX_MERGE_CONFLICT_RETRIES {
+                tracing::error!(
+                    task_id = task.id.0,
+                    retries,
+                    "merge conflict retry limit reached"
+                );
+                task_manager
+                    .update_task_status(&task.id, Status::Blocked)
+                    .await?;
+                let comment = format!("Auto-merge failed after {} rebase attempts: {}", retries, e);
+                let footer = attribution_footer("Commented", review_agent, review_model);
+                let _ = gh
+                    .add_comment(
+                        repo,
+                        &pr_number.to_string(),
+                        &format!("{}{}", comment, footer),
+                    )
+                    .await;
+                return Ok(());
+            }
+
+            // Try rebase in the worktree
+            let worktree_path =
+                engine::cleanup::store_get_field(store, repo, &task.id.0, "worktree").await;
+            if let Some(wt) = worktree_path {
+                let wt_path = std::path::PathBuf::from(&wt);
+                if wt_path.exists() {
+                    tracing::info!(
+                        task_id = task.id.0,
+                        worktree = %wt,
+                        "attempting rebase to resolve merge conflict"
+                    );
+                    let default_branch =
+                        config::get("gh.default_branch").unwrap_or_else(|_| "main".to_string());
+                    let rebase_result = tokio::process::Command::new("sh")
+                        .arg("-c")
+                        .arg(format!(
+                            "cd '{}' && git fetch origin && git -c commit.gpgsign=false rebase origin/{default_branch} && git push --force-with-lease",
+                            wt
+                        ))
+                        .output()
+                        .await;
+
+                    match rebase_result {
+                        Ok(out) if out.status.success() => {
+                            tracing::info!(
+                                task_id = task.id.0,
+                                "rebase succeeded — retrying merge"
+                            );
+                            store_increment(
+                                &Some(Arc::clone(store)),
+                                repo,
+                                &task.id.0,
+                                "merge_conflict_retries",
+                            )
+                            .await;
+                            if let Err(merge_err) = gh.merge_pr(repo, pr_number, true).await {
+                                tracing::error!(
+                                    task_id = task.id.0,
+                                    error = %merge_err,
+                                    "merge still failed after rebase — blocking"
+                                );
+                                task_manager
+                                    .update_task_status(&task.id, Status::Blocked)
+                                    .await?;
+                                return Ok(());
+                            }
+                            // Merge succeeded after rebase — fall through to done
+                            task_manager
+                                .update_task_status(&task.id, Status::Done)
+                                .await?;
+                            if let Err(ce) = cleanup_task_worktree(&task.id.0, repo, store).await {
+                                tracing::warn!(
+                                    task_id = task.id.0,
+                                    err = %ce,
+                                    "post-merge cleanup failed"
+                                );
+                            }
+                            return Ok(());
+                        }
+                        Ok(out) => {
+                            let stderr = String::from_utf8_lossy(&out.stderr);
+                            tracing::error!(
+                                task_id = task.id.0,
+                                stderr = %stderr,
+                                "rebase failed — blocking for human review"
+                            );
+                        }
+                        Err(io_err) => {
+                            tracing::error!(
+                                task_id = task.id.0,
+                                error = %io_err,
+                                "rebase command error — blocking for human review"
+                            );
+                        }
+                    }
+                }
+            }
+
+            // Rebase failed or no worktree — block
+            task_manager
+                .update_task_status(&task.id, Status::Blocked)
+                .await?;
+            let comment = format!(
+                "Auto-merge failed (merge conflict, rebase unsuccessful): {}",
+                e
+            );
+            let footer = attribution_footer("Commented", review_agent, review_model);
+            let _ = gh
+                .add_comment(
+                    repo,
+                    &pr_number.to_string(),
+                    &format!("{}{}", comment, footer),
+                )
+                .await;
+            return Ok(());
+        }
+
+        // Non-conflict merge failure (permissions, branch protection, etc.)
+        tracing::error!(task_id = task.id.0, error = %e, "merge failed — blocking for human review");
+        task_manager
+            .update_task_status(&task.id, Status::Blocked)
+            .await?;
+        let comment = format!("Auto-merge failed: {}", e);
+        let footer = attribution_footer("Commented", review_agent, review_model);
+        let _ = gh
+            .add_comment(
+                repo,
+                &pr_number.to_string(),
+                &format!("{}{}", comment, footer),
+            )
+            .await;
+        return Ok(());
+    }
+
+    // 6. Update status to done
+    store_set(
+        &Some(Arc::clone(store)),
+        repo,
+        &task.id.0,
+        &[("ci_merge_failures", serde_json::json!(0))],
+    )
+    .await;
+    task_manager
+        .update_task_status(&task.id, Status::Done)
+        .await?;
+
+    // 7. Cleanup worktree + branches
+    if let Err(e) = cleanup_task_worktree(&task.id.0, repo, store).await {
+        tracing::warn!(task_id = task.id.0, err = %e, "post-merge cleanup failed");
+    }
+
+    // 8. Post final comment on the PR
+    let comment = "✅ PR reviewed, approved, and merged.";
+    let footer = attribution_footer("Reviewed", review_agent, review_model);
+    let _ = gh
+        .add_comment(
+            repo,
+            &pr_number.to_string(),
+            &format!("{}{}", comment, footer),
+        )
+        .await;
+
+    tracing::info!(task_id = task.id.0, "auto-merge completed");
+
+    Ok(())
+}
+
+/// Handle review changes request — re-dispatch the original agent.
+#[allow(clippy::too_many_arguments)]
+pub(crate) async fn handle_review_changes(
+    task: &ExternalTask,
+    notes: &str,
+    issues: &[crate::engine::runner::response::ReviewIssue],
+    _backend: &Arc<dyn ExternalBackend>,
+    repo: &str,
+    pr_number: u64,
+    review_agent: &str,
+    review_model: &str,
+    task_manager: &Arc<TaskManager>,
+    store: &Arc<TaskStore>,
+) -> anyhow::Result<()> {
+    // 1. Check review cycle count (max 2 review rounds)
+    let review_cycles: u32 =
+        engine::cleanup::store_get_field(store, repo, &task.id.0, "review_cycles")
+            .await
+            .and_then(|v| v.parse().ok())
+            .unwrap_or(0);
+
+    let max_cycles: u32 = config::get("workflow.max_review_cycles")
+        .ok()
+        .and_then(|v| v.parse().ok())
+        .unwrap_or(2);
+
+    let gh = GhHttp::new()?;
+    let pr_num_str = pr_number.to_string();
+
+    if review_cycles >= max_cycles {
+        tracing::warn!(
+            task_id = task.id.0,
+            review_cycles,
+            max_cycles,
+            "max review cycles exceeded, blocking for human review"
+        );
+        task_manager
+            .update_task_status(&task.id, Status::Blocked)
+            .await?;
+        let escalation = format!(
+            "🔍 Review agent requested changes after {} cycles. Escalating to human.\n\n**Review Notes:**\n{}",
+            review_cycles, notes
+        );
+        let footer = attribution_footer("Commented", review_agent, review_model);
+        if let Err(e) = gh
+            .add_comment(repo, &pr_num_str, &format!("{}{}", escalation, footer))
+            .await
+        {
+            tracing::warn!(task_id = task.id.0, pr_number, err = %e, "failed to post escalation comment on PR");
+        }
+        return Ok(());
+    }
+
+    // 2. Build review context for the re-dispatched agent's prompt.
+    let mut comment = format!(
+        "🔍 Review agent requested changes (cycle {} of {}):\n\n{}",
+        review_cycles + 1,
+        max_cycles,
+        notes
+    );
+
+    if !issues.is_empty() {
+        comment.push_str("\n\n**Issues Found:**\n");
+        for issue in issues {
+            comment.push_str(&format!(
+                "- `{}` line {}: {} [{}]\n",
+                issue.file,
+                issue
+                    .line
+                    .map(|l| l.to_string())
+                    .unwrap_or_else(|| "?".to_string()),
+                issue.description,
+                issue.severity
+            ));
+        }
+    }
+
+    // 3. Store review context.
+    store_set(
+        &Some(Arc::clone(store)),
+        repo,
+        &task.id.0,
+        &[
+            (
+                "review_cycles",
+                serde_json::json!((review_cycles + 1) as i64),
+            ),
+            ("pr_review_context", serde_json::json!(comment.clone())),
+        ],
+    )
+    .await;
+
+    // 4. Set Routed to bypass LLM re-classification, reusing existing agent/model.
+    task_manager
+        .update_task_status(&task.id, Status::Routed)
+        .await?;
+
+    tracing::info!(
+        task_id = task.id.0,
+        review_cycles = review_cycles + 1,
+        pr_number,
+        "re-dispatching task (skipping re-route) to address review feedback on same PR"
+    );
+
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::backends::{ExternalId, ExternalTask, Mention, Status};
+    use crate::github::types::{GitHubReview, GitHubUser};
+
+    fn make_review(login: &str, state: &str, submitted_at: &str) -> GitHubReview {
+        GitHubReview {
+            id: 1,
+            user: GitHubUser {
+                login: login.to_string(),
+            },
+            body: None,
+            state: state.to_string(),
+            html_url: None,
+            submitted_at: submitted_at.to_string(),
+            commit_id: None,
+        }
+    }
+
+    #[test]
+    fn test_any_changes_requested_returns_true_when_latest_is_changes_requested() {
+        let reviews = vec![
+            make_review("alice", "APPROVED", "2026-01-01T10:00:00Z"),
+            make_review("alice", "CHANGES_REQUESTED", "2026-01-01T11:00:00Z"),
+        ];
+        assert!(
+            any_changes_requested_in_reviews(&reviews),
+            "should return true when reviewer's latest review is CHANGES_REQUESTED"
+        );
+    }
+
+    #[test]
+    fn test_any_changes_requested_returns_false_when_latest_is_approved() {
+        let reviews = vec![
+            make_review("alice", "CHANGES_REQUESTED", "2026-01-01T10:00:00Z"),
+            make_review("alice", "APPROVED", "2026-01-01T11:00:00Z"),
+        ];
+        assert!(
+            !any_changes_requested_in_reviews(&reviews),
+            "should return false when reviewer's latest review is APPROVED"
+        );
+    }
+
+    #[test]
+    fn test_any_changes_requested_returns_false_for_empty_reviews() {
+        assert!(
+            !any_changes_requested_in_reviews(&[]),
+            "should return false for empty review list"
+        );
+    }
+
+    #[test]
+    fn test_any_changes_requested_ignores_commented_state() {
+        let reviews = vec![make_review("alice", "COMMENTED", "2026-01-01T10:00:00Z")];
+        assert!(
+            !any_changes_requested_in_reviews(&reviews),
+            "COMMENTED state should be ignored"
+        );
+    }
+
+    #[test]
+    fn test_any_changes_requested_one_reviewer_blocks_even_if_other_approved() {
+        let reviews = vec![
+            make_review("alice", "APPROVED", "2026-01-01T10:00:00Z"),
+            make_review("bob", "CHANGES_REQUESTED", "2026-01-01T10:00:00Z"),
+        ];
+        assert!(
+            any_changes_requested_in_reviews(&reviews),
+            "should return true if any reviewer has outstanding CHANGES_REQUESTED"
+        );
+    }
+
+    /// Regression test: handle_review_changes must set status to Routed (not New).
+    #[tokio::test]
+    async fn handle_review_changes_sets_routed_not_new() {
+        use crate::engine::tasks::TaskManager;
+        use crate::store::{TaskStatus, TaskStore};
+        use async_trait::async_trait;
+
+        struct NoopBackend;
+        #[async_trait]
+        impl crate::backends::ExternalBackend for NoopBackend {
+            fn name(&self) -> &str {
+                "noop"
+            }
+            async fn create_task(
+                &self,
+                _t: &str,
+                _b: &str,
+                _l: &[String],
+            ) -> anyhow::Result<ExternalId> {
+                Ok(ExternalId("new".into()))
+            }
+            async fn get_task(&self, id: &ExternalId) -> anyhow::Result<ExternalTask> {
+                Ok(ExternalTask {
+                    id: id.clone(),
+                    title: "t".into(),
+                    body: "".into(),
+                    state: "open".into(),
+                    labels: vec![],
+                    author: "bot".into(),
+                    created_at: "2026-01-01T00:00:00Z".into(),
+                    updated_at: "2026-01-01T00:00:00Z".into(),
+                    url: "".into(),
+                })
+            }
+            async fn list_by_status(&self, _s: Status) -> anyhow::Result<Vec<ExternalTask>> {
+                Ok(vec![])
+            }
+            async fn list_routable(&self) -> anyhow::Result<Vec<ExternalTask>> {
+                Ok(vec![])
+            }
+            async fn post_comment(&self, _id: &ExternalId, _b: &str) -> anyhow::Result<()> {
+                Ok(())
+            }
+            async fn set_labels(&self, _id: &ExternalId, _l: &[String]) -> anyhow::Result<()> {
+                Ok(())
+            }
+            async fn remove_label(&self, _id: &ExternalId, _l: &str) -> anyhow::Result<()> {
+                Ok(())
+            }
+            async fn get_sub_issues(&self, _id: &ExternalId) -> anyhow::Result<Vec<ExternalId>> {
+                Ok(vec![])
+            }
+            async fn create_sub_task(
+                &self,
+                _p: &ExternalId,
+                _t: &str,
+                _b: &str,
+                _l: &[String],
+            ) -> anyhow::Result<ExternalId> {
+                Ok(ExternalId("child".into()))
+            }
+            async fn ensure_status_label(&self, _l: &str) -> anyhow::Result<()> {
+                Ok(())
+            }
+            async fn has_open_issue_with_title(&self, _t: &str, _l: &str) -> anyhow::Result<bool> {
+                Ok(false)
+            }
+            async fn health_check(&self) -> anyhow::Result<()> {
+                Ok(())
+            }
+            async fn is_pr_merged(&self, _b: &str) -> anyhow::Result<bool> {
+                Ok(false)
+            }
+            async fn get_authenticated_user(&self) -> anyhow::Result<Option<String>> {
+                Ok(Some("bot".into()))
+            }
+            async fn get_mentions(&self, _s: &str) -> anyhow::Result<Vec<Mention>> {
+                Ok(vec![])
+            }
+            async fn update_status(&self, _id: &ExternalId, _s: Status) -> anyhow::Result<()> {
+                Ok(())
+            }
+        }
+
+        let store = Arc::new(TaskStore::open_memory().await.unwrap());
+        let repo = "owner/repo";
+
+        let task_id_num = store
+            .create(&crate::store::NewTask {
+                external_id: None,
+                repo: repo.to_string(),
+                origin: "internal".to_string(),
+                title: "Implement feature X".to_string(),
+                body: "body".to_string(),
+                source: "cron".to_string(),
+                source_id: "daily".to_string(),
+                author: "".to_string(),
+                url: "".to_string(),
+                labels: vec![],
+            })
+            .await
+            .unwrap();
+        store
+            .update_status(task_id_num, crate::store::TaskStatus::InReview)
+            .await
+            .unwrap();
+
+        let task_id_str = format!("internal:{task_id_num}");
+        let backend: Arc<dyn crate::backends::ExternalBackend> = Arc::new(NoopBackend);
+        let task_manager = Arc::new(TaskManager::with_store(
+            Arc::clone(&backend),
+            Arc::clone(&store),
+            repo.to_string(),
+        ));
+
+        let task = ExternalTask {
+            id: ExternalId(task_id_str.clone()),
+            title: "Implement feature X".to_string(),
+            body: "body".to_string(),
+            state: "open".to_string(),
+            labels: vec![],
+            author: "".to_string(),
+            created_at: "2026-01-01T00:00:00Z".to_string(),
+            updated_at: "2026-01-01T00:00:00Z".to_string(),
+            url: "".to_string(),
+        };
+
+        let result = handle_review_changes(
+            &task,
+            "Please fix the type error on line 42",
+            &[],
+            &backend,
+            repo,
+            101,
+            "claude",
+            "sonnet",
+            &task_manager,
+            &store,
+        )
+        .await;
+
+        assert!(
+            result.is_ok(),
+            "handle_review_changes returned Err: {result:?}"
+        );
+
+        let updated = store.get(task_id_num).await.unwrap();
+        assert_eq!(
+            updated.status,
+            TaskStatus::Routed,
+            "handle_review_changes must set Routed, not New"
+        );
+
+        assert!(
+            !updated.pr_review_context.is_empty(),
+            "pr_review_context must be stored so the re-dispatched agent sees the feedback"
+        );
+        assert!(
+            !updated
+                .pr_review_context
+                .starts_with("A reviewer has requested changes"),
+            "pr_review_context must NOT start with the template header"
+        );
+        assert!(
+            updated
+                .pr_review_context
+                .contains("Please fix the type error on line 42"),
+            "pr_review_context must contain the review notes passed to handle_review_changes"
+        );
+
+        assert_eq!(
+            updated.review_cycles, 1,
+            "review_cycles must be incremented on each review change request"
+        );
+    }
+}
